Improved Efficency

+++ bits_helpers/spec_template.sh
@@ -1,3 +1,10 @@
+#!/bin/bash
+# We don't build RPMs if we have requires.json and provides.json. We can just proceed with checking dependencies.
+if [ -f "$INSTALLROOT/etc/requires.json" ] && [ -f "$INSTALLROOT/etc/provides.json" ]; then
+  exit 0
+fi
+
+# Build system-provides RPM and extract provides.json and store in $WORK_DIR/provides.json from where it will be copied by into each package provides.json
 if [ ! -f "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/system-provides-1-1.$(uname -m).rpm" ]; then
   mkdir -p "$WORK_DIR/rpmbuild"/{BUILD,BUILDROOT,RPMS,SOURCES,SPECS,SRPMS}
   cp $WORK_DIR/SPECS/$ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION/system-provides.spec "$WORK_DIR/rpmbuild/SPECS/system-provides.spec"
@@ -5,6 +12,13 @@ if [ ! -f "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/system-provides-1-1.$(uname -m).r
     --define "_topdir $WORK_DIR/rpmbuild" \
     --define "_buildarch $(uname -m)" \
     "$WORK_DIR/rpmbuild/SPECS/system-provides.spec"
+  
+  rpm -qp --queryformat "%{NAME}\n[%{PROVIDES}\n]" "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/system-provides-1-1.$(uname -m).rpm" | \
+  jq -R -s '
+    split("\n")
+    | map(select(length > 0))
+    | { (.[0]): .[1:] }
+  ' > "$WORK_DIR/provides.json"
 fi
 
 if [ "$PKGNAME" != defaults-* ] && [ -f "$WORK_DIR/SPECS/$ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION/${PKGNAME}.spec" ]; then
@@ -14,7 +28,7 @@ if [ "$PKGNAME" != defaults-* ] && [ -f "$WORK_DIR/SPECS/$ARCHITECTURE/$PKGNAME/
   source "$WORK_DIR/$ARCHITECTURE/rpm/latest/etc/profile.d/init.sh" || true
   cp "$WORK_DIR/SPECS/$ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION/${PKGNAME}.spec" "$WORK_DIR/rpmbuild/SPECS/"
   requires=()
-  for f in $FULL_REQUIRES; do
+  for f in $REQUIRES; do
     if [[ "$f" == "defaults-"* ]]; then
       continue
     fi
@@ -37,31 +51,50 @@ if [ "$PKGNAME" != defaults-* ] && [ -f "$WORK_DIR/SPECS/$ARCHITECTURE/$PKGNAME/
     --define "name ${PKGNAME}_${PKGVERSION}_${PKGREVISION}_${PKGHASH}" \
     --define "pkgname ${PKGNAME}" \
     --define "arch $(uname -m)" \
-    --define "path $ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION" \
-    --define "workdir $WORK_DIR" \
+    --define "installroot $INSTALLROOT" \
     --define "requires $requires_str" \
     --define "_topdir $WORK_DIR/rpmbuild" \
     --define "buildroot $WORK_DIR/rpmbuild/BUILDROOT/${PKGNAME}" \
     "$WORK_DIR/rpmbuild/SPECS/${PKGNAME}.spec"
-  
-  requires+=(${PKGNAME}_${PKGVERSION}_${PKGREVISION}_${PKGHASH})
-  requires+=(system-provides)
-  rpm -qp --queryformat '%{NAME}\n[%{PROVIDES}\n]' /home/akbehera/Desktop/repositories/rpm_1/rpmbuild/RPMS/x86_64/system-provides-1-1.x86_64.rpm
-  for f in "${requires[@]}"; do
-    rpm_file="${f}-1-1.$(uname -m).rpm"
-    rpm -qp --queryformat "%{NAME}\n[%{PROVIDES}\n]" \
-        "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/$rpm_file" | \
-        jq -R -s '
-            split("\n") | 
-            map(select(length > 0)) | 
-            { (.[0]): .[1:] }
-        '
-  done | jq -s 'add' > "$WORK_DIR/$ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION/provides.json"
 
   rpm -qp --queryformat "%{NAME}\n[%{REQUIRES}\n]" "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/${PKGNAME}_${PKGVERSION}_${PKGREVISION}_${PKGHASH}-1-1.$(uname -m).rpm" | \
   jq -R -s '
     split("\n")
     | map(select(length > 0))
     | { (.[0]): .[1:] }
-  ' > "$WORK_DIR/$ARCHITECTURE/$PKGNAME/$PKGVERSION-$PKGREVISION/requires.json"
+  ' > "$INSTALLROOT/etc/requires.json"
+
+  rpm -qp --queryformat "%{NAME}\n[%{PROVIDES}\n]" "$WORK_DIR/rpmbuild/RPMS/$(uname -m)/${PKGNAME}_${PKGVERSION}_${PKGREVISION}_${PKGHASH}-1-1.$(uname -m).rpm" | \
+  jq -R -s '
+    split("\n")
+    | map(select(length > 0))
+    | { (.[0]): .[1:] }
+  ' > "$INSTALLROOT/etc/provides.json"
+
+  # Collect all provides.json for all the package mentioned in $REQUIRES and also system-provides.json and create a single provides.json file from where we can see if each
+  # REQUIRES is satisfied or not.
+  provides_files=("$INSTALLROOT/etc/provides.json")
+  provides_files+=("$WORK_DIR/provides.json")
+
+  for f in $REQUIRES; do
+    if [[ "$f" == "defaults-"* ]]; then
+      continue
+    fi
+    F=${f^^}
+    F=${F//-/_}
+    hash="${F}_HASH"
+    ver="${F}_VERSION"
+    rev="${F}_REVISION"
+    
+    provides_file="${WORK_DIR}/$ARCHITECTURE/$f/${!ver}-${!rev}/etc/provides.json"
+    
+    if [ -f "$provides_file" ]; then
+      provides_files+=("$provides_file")
+    fi
+  done
+
+  # Merge all provides.json files into a single file and store it in $INSTALLROOT/etc/provides.json
+  jq -s 'reduce .[] as $item ({}; . * $item)' "${provides_files[@]}" > "$INSTALLROOT/provides.json"
+  rm $INSTALLROOT/etc/provides.json
+  mv $INSTALLROOT/provides.json $INSTALLROOT/etc/provides.json
 fi
